requirements.txt
pillow
screeninfo